Refactor compose components
